Move credential support into creds
