+++ v2/internal/testcommon/creds/mock_token_cred.go
@@ -3,7 +3,7 @@ Copyright (c) Microsoft Corporation.
 Licensed under the MIT license.
 */
 
-package testcommon
+package creds
 
 import (
 	"context"
